Update tasks list after task creation
