+++ angular-project/package.json
@@ -10,11 +10,11 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.3",
-    "@angular/cdk": "^17.1.2",
+    "@angular/animations": "^17.2.2",
+    "@angular/cdk": "^17.3.7",
     "@angular/common": "^17.1.3",
     "@angular/compiler": "^17.1.3",
-    "@angular/core": "^17.3.7",
+    "@angular/core": "^17.3.10",
     "@angular/forms": "^17.1.3",
     "@angular/material": "^17.1.2",
     "@angular/platform-browser": "^17.1.3",
