+++ angular-project/src/app/tasklist/task.service.ts
@@ -2,6 +2,7 @@ import { Injectable, Inject, signal } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { ActivatedRoute } from '@angular/router';
 import { TaskDetailDTO } from '../taskdetail/TaskDetailDTO';
+import { TasksDTO } from './task';
 
 
 @Injectable({
@@ -10,10 +11,11 @@ import { TaskDetailDTO } from '../taskdetail/TaskDetailDTO';
 
 export class TaskService {
   constructor(private http: HttpClient, @Inject('BASE_URL') private baseUrl: string, private route: ActivatedRoute) { }
+
   createTask(Task: CreateTaskDTO) {
-    return this.http.put<CreateTaskDTO>(this.baseUrl + '/tasklist', Task);
+    return this.http.put<TasksDTO>(this.baseUrl + '/tasklist', Task);
   };
-  
+
   getTaskDetails(id: number){
     return this.http.get<TaskDetailDTO>(this.baseUrl + "/taskdetail/" + id);
   }
@@ -43,6 +45,6 @@ export interface CreateTaskDTO {
   taskName: string;
   taskDescription: string;
   taskPriority: number;
-  taskDeadline: Date; 
+  taskDeadline: Date;
 }
