+++ angular-project/src/app/tasklist/task.ts
@@ -1,0 +1,8 @@
+export interface TasksDTO {
+  id: number;
+  name: string;
+  description: string;
+  priority: number;
+  isDone: boolean;
+  deadline: Date;
+}
