+++ angular-project/src/app/tasklist/tasklist.component.ts
@@ -1,18 +1,15 @@
 import { DatePipe, NgFor, NgIf, CommonModule} from '@angular/common';
 import { HttpClient } from '@angular/common/http';
-import { Component, Inject, signal, Injectable, AfterViewInit, OnInit, OnDestroy } from '@angular/core';
-import { FormControl, FormGroup, ReactiveFormsModule, FormsModule, Validators, FormBuilder } from '@angular/forms';
-import { TaskService, CreateTaskDTO } from './task.service';
-import { Router, RouterModule, ActivatedRoute, RouterLink } from '@angular/router';
-import { Subject, takeUntil } from 'rxjs';
-import { MatDatepickerModule } from '@angular/material/datepicker';
+import { Component, Inject, signal, AfterViewInit, DestroyRef, inject } from '@angular/core';
+import { FormControl, FormGroup, ReactiveFormsModule, FormsModule, Validators } from '@angular/forms';
+import { TaskService } from './task.service';
+import { RouterLink } from '@angular/router';
 import { TaskDetailDTO } from '../taskdetail/TaskDetailDTO';
-import { LiveAnnouncer } from '@angular/cdk/a11y';
-import { ViewChild } from '@angular/core';
-import { MatSort, Sort, MatSortModule } from '@angular/material/sort';
-import { MatTableDataSource, MatTableModule } from '@angular/material/table';
+import { MatSortModule } from '@angular/material/sort';
+import { MatTableModule } from '@angular/material/table';
 import { SearchFilterPipe } from 'src/app/tasklist/search-filter.pipe';
-
+import { TasksDTO } from './task';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 @Component({
   selector: 'app-tasklist',
@@ -24,17 +21,24 @@ import { SearchFilterPipe } from 'src/app/tasklist/search-filter.pipe';
   styles: []
 })
 export class TasklistComponent implements AfterViewInit {
-
-searchTerm: string;
+  destroyRef = inject(DestroyRef);
+  searchTerm: string;
+  taskData = signal<TasksDTO[]>([]);
+  xd: boolean = false;
+  taskForm = new FormGroup(
+    {
+      name: new FormControl('', Validators.required),
+      description: new FormControl('', Validators.required),
+      priority: new FormControl(null, Validators.required),
+      deadline: new FormControl(null, Validators.required),
+    })
 
   constructor(
-    private route: ActivatedRoute, http: HttpClient,
-    private _liveAnnouncer: LiveAnnouncer,
-    private router: Router,
+    http: HttpClient,
     private taskService: TaskService,
 
     @Inject("BASE_URL") baseUrl: string) {
-    http.get<TasksDTO[]>(baseUrl + '/tasks').subscribe(result => { this.TaskData.set(result) }, error => console.error(error));
+    http.get<TasksDTO[]>(baseUrl + '/tasks').subscribe(result => { this.taskData.set(result) }, error => console.error(error));
   }
   deletni = signal<TaskDetailDTO>(undefined);
 
@@ -59,29 +63,6 @@ searchTerm: string;
 
   }
 
-  taskName: string = "no data";
-  taskDescription: string = "no data";
-  taskPriority: number = 0;
-  taskDeadline: Date;
-  private unsubscribe$ = new Subject<void>();
-
-
-  taskForm = new FormGroup(
-    {
-      name: new FormControl('', Validators.required),
-      description: new FormControl('', Validators.required),
-      priority: new FormControl(null, Validators.required),
-      deadline: new FormControl(null, Validators.required),
-    })
-
-
-
-  TaskData = signal<TasksDTO[]>([]);
-  // newTask = signal<TasksDTO>(undefined);
-  taskINFO = signal<CreateTaskDTO>(undefined);
-  xd: boolean = false;
-  taskIdFromRoute: number = 0;
-  taskDetailInfo = signal<TaskDetailDTO>(undefined);
   onAddTask() {
     if (this.taskForm.valid) {
       this.taskService.createTask({
@@ -89,15 +70,12 @@ searchTerm: string;
         taskDescription: this.taskForm.controls['description'].value,
         taskPriority: this.taskForm.controls['priority'].value,
         taskDeadline: this.taskForm.controls['deadline'].value,
-      }).subscribe(TaskINFO => this.taskINFO.set(TaskINFO), this.router.navigate['/tasklist']);
+      }).pipe(takeUntilDestroyed(this.destroyRef))
+        .subscribe(taskInfo => this.taskData.update(tasks => [...tasks, taskInfo]));
     }
   }
 
-  private dataSubject = new Subject<[TaskDetailDTO]>();
-  data$ = this.dataSubject.asObservable();
-
   onDelete(id: number) {
-    
     this.taskService.deleteTask(id).subscribe();
   }
 
@@ -114,25 +92,6 @@ searchTerm: string;
       elem.innerHTML = "Show finished tasks";
       this.xd = false;
     }
-          }
-
-
-  ngOnInit(): void {
-    const RouteParams = this.route.snapshot.paramMap;
-    this.taskIdFromRoute = Number(RouteParams.get('id'));
-    console.log(RouteParams);
-    this.taskService.getTaskDetails(this.taskIdFromRoute).pipe(takeUntil(this.unsubscribe$)).subscribe(taskDetail => { this.taskDetailInfo.set(taskDetail); });
   }
-
-
-}
-
-export interface TasksDTO {
-  id: number;
-  name: string;
-  description: string;
-  priority: number;
-  isDone: boolean;
-  deadline: Date;
 }
